+++ crates/engine/src/plugin/wasm/manager.rs
@@ -1,7 +1,7 @@
 use super::HandlerError;
 use crate::error;
 use crate::plugin::{wasm::Plugin, Handler, Runtime};
-use common::GameState;
+use common::{Event, GameState};
 use std::{fmt, fs, path::Path};
 use wasmtime::Store;
 
@@ -25,9 +25,13 @@ impl fmt::Debug for Manager {
 }
 
 impl Handler for Manager {
-    fn run_plugins(&mut self, game_state: &mut GameState) -> Result<(), error::Runtime> {
+    fn run_plugins(
+        &mut self,
+        game_state: &mut GameState,
+        events: &[Event],
+    ) -> Result<(), error::Runtime> {
         for plugin in &mut self.plugins {
-            plugin.run(game_state)?;
+            plugin.run(game_state, events)?;
         }
 
         Ok(())
@@ -71,7 +75,7 @@ mod tests {
             let mut game_state = GameState::default();
             let mut manager = Manager::default();
 
-            assert!(manager.run_plugins(&mut game_state).is_ok())
+            assert!(manager.run_plugins(&mut game_state, &[]).is_ok())
         }
 
         #[test]
@@ -86,7 +90,7 @@ mod tests {
             let p = plugin(WAT_VALID);
             manager.plugins.push(p);
 
-            assert!(manager.run_plugins(&mut game_state).is_ok())
+            assert!(manager.run_plugins(&mut game_state, &[]).is_ok())
         }
 
         #[test]
@@ -101,7 +105,7 @@ mod tests {
             let p = plugin(WAT_MISSING_FUNC);
             manager.plugins.push(p);
 
-            let err = anyhow::Error::new(manager.run_plugins(&mut game_state).unwrap_err());
+            let err = anyhow::Error::new(manager.run_plugins(&mut game_state, &[]).unwrap_err());
 
             assert_eq!(
                 format!("{:?}", err),