+++ crates/sdk/src/prelude.rs
@@ -1,5 +1,5 @@
 pub use crate::Sdk;
 pub use anyhow::{self, bail, Result};
 pub use common::{
-    serde_json, Deserialize, PluginState, Registration, Serialize, StateTransfer, Value,
+    serde_json, Deserialize, Event, Key, PluginState, Registration, Serialize, StateTransfer, Value,
 };